feat: adiciona tratamento de excecoes e finaliza CRUD de fornecedor

+++ src/main/java/com/haackdev/commercial_management/service/FornecedorService.java
@@ -2,7 +2,11 @@ package com.haackdev.commercial_management.service;
 
 import com.haackdev.commercial_management.entity.Fornecedor;
 import com.haackdev.commercial_management.repository.FornecedorRepository;
+import com.haackdev.commercial_management.service.exceptions.DatabaseException;
+import com.haackdev.commercial_management.service.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -20,7 +24,8 @@ public class FornecedorService {
 
     // Busca um fornecedor pelo ID
     public Fornecedor findById(Long id) {
-        return fornecedorRepository.findById(id).orElse(null);
+        return fornecedorRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
     // Insere um novo fornecedor
@@ -30,14 +35,25 @@ public class FornecedorService {
 
     // Deleta um fornecedor pelo ID
     public void delete(Long id) {
-        fornecedorRepository.deleteById(id);
+        if (!fornecedorRepository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+        try {
+            fornecedorRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Não é possível deletar um fornecedor que possui produtos cadastrados");
+        }
     }
 
     // Atualiza um fornecedor existente
     public Fornecedor update(Long id, Fornecedor fornecedor) {
-        Fornecedor entity = fornecedorRepository.getReferenceById(id);
-        updateData(entity, fornecedor);
-        return fornecedorRepository.save(entity);
+        try {
+            Fornecedor entity = fornecedorRepository.getReferenceById(id);
+            updateData(entity, fornecedor);
+            return fornecedorRepository.save(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException(id);
+        }
     }
 
     // Metodo auxiliar para atualizar os dados do fornecedor
